Instance ArrayObject objects with forthcoming variant
Note that, at the moment, Dict and Array instances of
Instances and Disks are only used in tests. So, when
adding those instances, we do not have to worry
about backwards compatibility. We add the "forthcoming"
field in front of the Array representation; in this way,
if new fields will be added later, the old representation
will be a prefix of the then new one.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Niklas Hambuechen <niklash@google.com>