Adapt instance existence check to new locking
With the change to the new locking system, there is no
need any more to create locks at run time. As a consequence
of this, no check happens at "lock addition" that the lock
doesn't exist already. Fix the old comment still claiming this
and repeat the instance existence check at CheckPrereq.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>