Use disk based export summary
The backend writes a file summarizing the instance including the disk
template after export. Without the disk template of the instance, this
is no longer possible. However since it might be nice to be able to
import an instance into a pre-disk-template-removal cluster, the disk
template will still be provided (but might be unexpectedly 'mixed').

Signed-off-by: Aaron Karper <akarper@google.com>
Reviewed-by: Michele Tartara <mtartara@google.com>