kvm: New _with_qmp decorator
The decorator takes the hypervisor and the instance object as passed
to all hotplug related commands and ensures a valid QmpConnection()
object is found in hypervisor's qmp attribute.

This decorator is responsible for all actions related to the
QmpConnection (e.g. init, connect, close) and for proper exception
propagation. In order to be safe against nested invocations we close
the connection only if we create it.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>
Reviewed-by: Hrvoje Ribicic <riba@google.com>