qa: Re-enable NIC hotplug tests
This was disabled due to Issue 885. This issue states that NIC
hotplugging sometimes makes an instance unresponsive. This cannot be
the case, since NIC hotplug takes place on an extra NIC and does not
mess up with instance's primary NIC. Still this sometimes might be
related with --online fix (see commit b0a383a).

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>
Reviewed-by: Hrvoje Ribicic <riba@google.com>