TestCommon: Fix QuickCheck import warnings
This only appears on systems with QuickCheck >= 2.7.

For TestCommon, it happens because the QC qualified name is only used
in the conditional section.
Fixed by making the import conditional as well.

For Statistics, the `Test.Ganeti.TestCommon` import was not necessary
for QC 2.7 because there `Test.QuickCheck` already provides `counterexample`.
Fixed by giving an import list for `Test.QuickCheck`.

Signed-off-by: Niklas Hambuechen <niklash@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>