Don't check certificates if we check the content
When building the schroot, tell wget not to verify certificates.
It is not necessary, as we verify the SHA1 hash of the downloaded
content anyway. Moreover, it doesn't even help, as the only way
we end up at a https page is by getting redirected there from
a normal http query.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>