Make clutter test optional
The previous patch introduced a check to
'gnt-cluster verify' which verifies whether ganeti
cluttered up the 'authorized_keys' file. This is possible
to happen if cluster operations are aborted mid-way or
finish unsuccessfully. However, as Ganeti is not the sole
authority on the 'authorized_keys' file, we cannot assume
that no other systems added keys. Leaving the clutter
check non-optional would cause too many false positives
for many setups. Therefore, we add an option
'--verify-ssh-clutter' to 'gnt-cluster verify'.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>