Replace and Recreate disks without using NAL
These operations may run the IAllocator to decide on
where to create a new disk and hence ask for a lot
of node locks. Therefore, it was necessary to ask for
the node allocation lock in order to protect jobs
using opportunistic locking from running out of
nodes. This necessity does not exist any more, hence
don't ask for the node allocation lock either.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Niklas Hambuechen <niklash@google.com>