Deal with nonexisting public key file
After an upgrade or if the cluster is in some inconsistant
state, it can happen that the public key file does not
exist. In this case the cluster verify symptoms were rather
cryptic. Fix this by creating the file on the fly if it
does not exist and emit a useful error message.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>