htools: make checkData idempotent
The usual understanding of a check function is that
it does not modify the underlying data as to achieve
an inconsistent state. However, by the unfortunate
combination of checking (and fixing xMem) and "reserving"
memory for the down instances by reducing the amount
of free memory on the node that property was violated.

Make checkData again a pure check (that only fixes
values and hence behaves in an idempotent way). Memory
needs for the down instances will be taken into account
together with the memory usage of the forthcoming instances.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Signed-off-by: Niklas Hambuechen <niklash@google.com>