Rename lxc_tty to num_ttys
As noted in the previous commit, the lxc_ prefix is bad and it's good
to be rid of it, and we can take advantage of this change to rename the
variable.
Its previous name was meant to reflect the lxc config option being set
(lxc.tty), but we should use a more descriptive name.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>