HTools: Node violation checks for forthcoming instances
This implements basic support for forthcoming instances in htools,
in particular whether a forthcoming instance to be placed fits on it
and the maintainance of the relevant statistics to decide this.

Each statistic in the `Node` data type that has to handle forthcoming
instances gets a `somethingForth` equivalent that includes the
corresponding `something` statistic for real instances.

For example, when `fMem` measures the free memory taking all real
instances into account, `fMemForth` is the free memory on the node
in case all forthcoming instances on the node were also made real,
thus being less or equal than `fMem` at any given time.

This commit does NOT yet
- handle the maintenance of rMemForth, failN1Forth, and peersForth,
  so n+1 safety for forthcoming instances is not yet implemented
- add the new `somethingForth` fields to the cluster balancedness
  measure (`detailedCVInfoExt`)
- add `test/data/htools/*.json` or `*.data` files that test this
  feature

All unit tests pass.

Signed-off-by: Niklas Hambuechen <niklash@google.com>