Allow for checking of individual cgroup parameter presence
To alter our behavior depending on whether the kernel used supports
certain cgroup parameters or not, this patch adds a helper function
capable of checking that. The function is heavily based on earlier
functions for reading cgroup parameter values, with some refactoring
to keep the code in a readable state.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>