Make usage of memory.memsw.limit_in_bytes optional
In the lxc code, it used to be that whenever the amount of memory the
container has is changed via the memory.limit_in_bytes parameter, the
memory.memsw.limit_in_bytes parameter had to be changed as well. The
latter is memory+swap, and not changing it results in fun behavior.

The presence of the memsw parameter is dependent on some kernel
parameters, and in some cases we might crash while trying to set it.
This patch sets it only if present, allowing us to work on more
configurations.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>