Gluster: fix mount.glusterfs argument passing
Currently, the Gluster server port is included in the volume path,
i.e., <host>:<port>:<volumeid>, but the mount.glusterfs script requires
to explicitly specify the server port using the '-o server-port=port'
option. This would cause the port setting to be ignored and a default of
24007 to be used when using glusterfs versions later than 3.3.0 with
Ganeti, while for newer versions glusterfs mount fails.

This patch modifies '_GetFUSEMountString' method to generate the correct
argument string for the mount.glusterfs script. Also, it updates the
'ganeti.storage.gluster_unittest.py' unit test to correspond to the new
changes.

This fixes issue #984.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>