Also consider forthcoming instances in cluster score
When computing the cluster score, also take into account
how unbalanced the cluster will be once all forthcoming
instances are realized. To be backwards compatible with
the old metric, this is done by splitting the coefficient
for those metric components over the two scenarios. Note
that if no forthcoming instances are present, the two
scenarios coincide.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>