Remove the rapi-workload test suite
The rapi-workload test suite was introduced in 2.10 to try and provide
a chunk of workload which could be used to detect regressions in the
RAPI functionality and format. It served its purpose, helping detect a
few important regressions in time, but with the assistance of manual
tools which were never automated. Automating them proved difficult, and
probably infeasible.

The flaws in the QA that caused the regressions were removed over time,
and with 2.13 around the corner, it's time to shoot this horse. This
patch removes the file, and a sole mention of it in the Makefile.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>