Also test that migration tags are identified correctly
The specification says, that if a cluster tag starts htools:migration:x
then node tags starting with x: indicate migration restrictions. Note
the additional colon. To test for this, add on a cluster with cluster
tag 'htools:migration:hv' a tag starting with 'hv' that is not a migration
tag.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>