Add check for >1 disk used with the LXC hypervisor
This patch adds a check that an LXC instance does not have more than
one disk. While this restriction may be lifted in the future, for the
time being it will stay in place. Rather than confuse the user, we
choose to report this as an error immediately.

It would be nice if we could do this before creating disks and other
resources, but Ganeti currently has no per-hypervisor checks of
instance configuration, and no need for them apart from this use case.
Until another use comes up, refactoring will probably be a bad idea,
especially as this restriction might be lifted in a later version.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>