Count common failures in instance location
Keep statistics about the number of pairs of instances and common-failure
tags where primary and secondary node of that instance share the failure
tag.

The data structure is organized as follows. Each instance knows the
number of common failures between its primary and its secondary node.
This value is updated by setInstanceLocationScore (this function cannot
be part of the Instance module as this would cause circular dependencies).
Each node keeps track of the sum of the location scores of its primary
instances and updates them with addPri and removePri. Note that this
requires the location score of an instance to be set before the instance
is placed on the node.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>