Add a test for location tags
In this example, the cluster is well balanced. Each node is
primary and secondary of precisely one instance. However,
the placement of the instances is sub optimal location wise.
All instances are located entirely in one failure domain,
even though it would be possible to have primary and secondary
of each instance in a different failure domain.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>