Ignore spurious pylint warnings
The function ExtStorageFromDisk returns a poor man's sum
of an ExtStorage object and a string, i.e., it returns
either the pair of True and the ExtStorage object, or
False and a human-readable error message. Pylint, however,
is not capable of understanding that concept, and, as in
one branch the second component of the pair is a string,
it assumes that it always is a string. This produces
spurious warnings later when the ExtStorage object is
used. Ignore those.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Aaron Karper <akarper@google.com>