Mention tags in --restricted-migration man page entry
The --restricted-migration option was added to allow evacuating nodes
in the middle of a hypervisor upgrade, where the only migrations you
were sure to be save were those off the evacuated node. In the mean
time, migration tags have been added for precisely the purpose of
hypervisor upgrades. So mention in the man pages that those usually
are the better option.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>