Add a function computing the test instances
The allocation vector (and hence the lost-allocations metrics build
on top of it) is based on counting how many instances of a certain
size can be placed on a node. Add a function computing those test
instances from the policy, i.e., add a returning an instance matching
the lower bounds of each policy interval, in decreasing order.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>