Add a function computing the lost-allocations metric
This is the difference between the allocation vector before and after
the instance is placed on the node, and the remaining disk space as
a tie breaker. Note that the natural order on ([Int], Int) is the order
of the lost-allocations metric.

As it might not be possible to place the instance on the node at all,
the metric is wrapped in an OpResult.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>