Restore previous snapshot behaviour for LVM
The "Make snapshotting for exports optional" commit refactored the use
of snapshots, allowing non-LVM disks to be exported as well. It
inadvertently introduced incompatible behaviour: snapshots used to be
made even if a live instance was to be shutdown during the export,
primarily to reduce instance downtime as it could be started
immediately after a snapshot was made. The patch tried to account for
that, but the old behaviour was fundamentally incompatible with the
simple solution for enabling exports of different disks.

This patch restores the previous LVM snapshot behavior, but breaks
exports of different disk templates with live instances. This is simply
because we have no means of identifying when a snapshot can be made,
and follow-up patches will address this. Exports of offline instances
will still work, as previous patches make it so that no snapshots are
required when an instance is offline.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>