Use UUID and not logical_id to mark exported disks
Ganeti used to use the logical_id of a disk to name the file containing
a locally exported disk. This worked while only plain or drbd disks
were used, but broke down with e.g. file-based disks. The reason is
simple - the logical_id is specific to each disk template, and in the
case of files, contains the full path to the file, causing various
issues when used.

This patch uses the newly introduced disk UUID instead of the
logical_id, adding it to snapshot generation where it was previously
missing. While the logical_id did have some meaning (Linux UUID), the
UUID was actually taken from an LVM snapshot and did not have any
meaning due to the short life of the snapshot. The id format also
differed between plain and drbd, which makes sense as it was used only
as a reference - much as it should for compatibility with e.g. OVF.
This also means we maintain backward compatibility with older Ganeti
versions.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>