Retry on more network errors
Retries were added to luxi/wconfd RPC calls in f3aebf6ff0
to deal with the race condition of a connect happening
at precisely the time an idle connection timed out. This,
however, did not cover the situation of a timeout in precisely
the moment the client is sending the request which will look
for both parties as if the other decided to not send any data,
resulting in a ConnectionClosedError. Thus, retry on those
errors as well.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>