Return success of snapshotting and refactor code
The ExportInstanceHelper used to have to be told in advance whether to
snapshot disks or not, and despite that had an additional separate
method that had to be invoked to snapshot disks. This patch changes the
code to retain just the method, and have it return whether snapshotting
was successful for future purposes.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>