Increase waiting time for WConfD
As the Ganeti 2.12 design allows for WConfD to be
restarted with jobs running, jobs accept WConfD not
be present and wait for to come back, retrying.
However, the time the jobs waited for WConfD to come
back was shorter than an actual restart might take.
So, increase it significantly and also back off with
the number of retries increasing to avoid flooding
WConfD after its restart.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>