Add a data collector for xentop
We already have a CPU load collector based on /proc/stat; however,
in a Xen setup this only measures the dom0 CPU load which is not
that useful, if we're interested in the activity of the instances
on that node. Therefore add a xen-specific CPU-load data collector,
based on xentop.

From xentop we can obtain the total amount of vCPU seconds used since
an arbitrary start point. We collect those observations together with
time stamps over the averaging period for CPU load. When asked for CPU
average, it can be easily computed from the first and last data point
kept.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>