Refactor optional mechanism for ExtStorage scripts
Up until now, the optional sense was related to provider validity.
gnt-storage diagnose would not report errors in case an optional
script was missing, however if the logic ended up running the
corresponding script, then Ganeti would fail loudly.

This patch changes the above logic which is error prone and not
extensible for future functionality. From now on, if an optional
script is missing then we assume this to be equivalent with a no-op
script. Ganeti will no longer fail, but instead print debug
messages.

Apply some fixes in the ganeti-extstorage-interface man page too.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>