Add the 'exclusive' parameter in bdev.Open()
Open makes the device ready for I/O. The exclusive argument
implies that the device will be used only by one and no concurrent
access will be made on this device. By default this is True.

Currently, the only case that we have concurrent access on the
device is during migrations.

This patch will allow for further unification with DRBD code.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>