Introduce blockdev_open RPC
Until now there was a blockdev_close RPC that was used exclusively
by DRBD code.

This patch adds the blockdev_open RPC which will allow for
further unification of DRBD and other templates code.

The RPC takes the disks of an instance and whether they should be
exclusively opened or not. Besides opening the block device in the
backend it creates the symlinks of the disks that the hypervisor
requires.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>