Add helpers for Open/Close during migration
Introduce _OpenInstanceDisks() and _CloseInstanceDisks() helpers
that will be used during migration. The first takes the exclusive
argument that denotes whether the disks will be shared or not. For
now we keep the separation between DTS_EXT_MIRROR and DTS_INT_MIRROR
code, but we introduce the entry points needed for the ExtStorage
interface that were missing.

The workflow during migration for DTS_EXT_MIRROR disk templates
will look like this:

------------------------------------------------------
         source           |          target
------------------------------------------------------
                          |      _CloseInstanceDisks()
_OpenInstanceDisks(SHA)   |
                          |    _OpenInstanceDisks(SHA)
------------------------------------------------------
                      MIGRATION
------------------------------------------------------
_CloseInstanceDisks()     |
                          |   _OpenInstanceDisks(EXCL)
------------------------------------------------------

If migration fails:

------------------------------------------------------
                          |      _CloseInstanceDisks()
_OpenInstanceDisks(EXCL)  |
------------------------------------------------------

In case of a migration cleanup:

------------------------------------------------------
        running           |           demoted
------------------------------------------------------
                          |      _CloseInstanceDisks()
_OpenInstanceDisks(EXCL)  |
------------------------------------------------------

With the introduction of _CloseInstanceDisks(), the BlockdevClose()
backend function is now used for all bdev types and not only for DRBD.
Fix the corresponding debug message.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>