Move common code out of DrbdAttachNet
DrbdAttachNet is used only by the _GoReconnect() helper during
migration. Up until now, in case of multimaster it created the
symlinks needed by the hypervisor and opened the devices so that
they could be ready for I/O.

With the introduction of _OpenInstanceDisks() we can now unify the
code and let DrbdAttachNet do only DRBD-specific actions.

Now DTS_INT_MIRROR only differs from DTS_EXT_MIRROR in the
following steps during migration:

 _GoStandAlone()
 _GoReconnect()
 _WaitUntilSync()

which are all DRBD-specific.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>