Add a constant for the xen cpu observation threshold
With xentop we can obtain CPU usage since start with an accuracy
of seconds. So when computing CPU load as difference between two
observations, the observations better be sufficiently apart for
the result to be meaningful. Specify a threshold for this; we
choose 100 seconds to have the observation inaccuracy below percent
of a CPU.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>