Gracefully handle CPU counter rollovers
The CPU-usage counter in xentop can roll over. In this case,
using the difference between old and new value would give a
negative usage, which is obviously wrong. To avoid this mistake,
we compare old and new usage before combining values. If we
observe a rollover, we discard this interval, but we keep
the old observations, appropriately moved.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>