Clean out old instances in Xen CPU load
If we don't find an instance in the output of xentop any
more that usually means it moved away. In this case, it might
be useful to still have the old CPU load as a best guess until
a new value is available on the new location. But after this
period, we should stop reporting usage information for that
instance.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>