Signal to the job queue when the cluster is gone
When destroying a cluster, the single-job job queue process tries to
load the job file, which is already gone for good, resulting in the
process being stuck forever.

This patch ensures that the job queue terminates in such a case.

The added static variable to LUClusterDestroy should be removed in the
future together with the Python job queue.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>