Use CPP macros to chose the appropriate NFData instance
We need the CollectorData to be an instance of NFData.
As that is build on Data.Sequence.Seq, it is natural to
just decent to the argument. This, however, requires
Data.Sequence.Seq to be an instance of NFData, which
it only became in version 0.5.0.0 of the containers
library. So keep the work-around for older library
versions, while using the natural definition where
available.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>