Fix issues when generating 'lens' version definitions
The original version didn't work when the version was just 2 numbers,
and also wasn't robust enough when there were multiple 'lens' versions
installed.

This patch fixes that, and always takes the last installed version, as
reported by ghc-pkg.

Starting with 2.14 there is a native cabal-based mechanism for
generating these definitions, so the whole 'hs-lens-version' target
should be dropped from the Makefile in that version.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>