Add a design document on improving allocation
This design describes an addition to the cluster
metrics that improves allocation efficiency. So
far, only the balancedness of the amount of reserved
memory was taken into account, but not its total
amount. The reserved memory, however, is not preserved
by instance moves; it can be lowered by spreading the
secondaries more widely over the cluster.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>