Change wording in documentation wrt configure-time paths
It is a deliberate choice to have the full generated documentation
part of the distribution, as this reduces the dependencies for users
who just want to install and use Ganeti. It has, however, the disadvantage
that we ship files with configure-time paths substituted in. So be
careful with the wording to indicate that those paths depend on
configure-time choices.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>