Support no-master state in ssconf
When destroying a cluster, it needs to go to no-master state
first, to ensure daemons will refuse to restart. So, even if
there is no master, we need to provide a value the master entry
in ssconf. The value chosen contains a space character and therefore
cannot be a valid domain name.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>