Make LuxiD clean up its lock file
LuxiD locks the job queue during its life time to avoid
two instances running in parallel. When the process dies
(for whatever reason) the advisory lock on the file is gone
so a new daemon can restart. However, the file itself is
not gone, so after cluster destruction a gnt-cluster init
will refuse to create a new cluster. So, after finishing,
remove the file as well.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>