Add default values for hv_state_static and disk_state_static
When parsing the cluster configuration, accept that "hv_state_static"
and "disk_state_static" can be missing; in this case, just use the
default value. This follows the implicit upgrade that was present
for these parameters in the python code all the time. In this way,
upgrading from very old versions of Ganeti runs more smoothly
(issue 1019).

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>