Make writeConfigAndUnlock retryable
We have the built-in assumption that daemons can be restarted
at any time. This includes situations where an action is already
taken, but the caller did not get the answer jet. For the case
of writeConfigAndUnlock this can lead to the command being reissued;
of course, it should not do anything on second attempt, but fail in
a way the caller can live with.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>