Increase number of retries for daemon RPCs
As the Ganeti 2.12 design allows for daemons to be
restarted with jobs running, jobs accept them not
to be present and wait for them to come back, retrying.
However, the total time jobs waited, in some conditions,
turned out to be shorter than the actual restart. So
increase the number of retries and also linearly back
off after failures.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>