Add a predicate testing if a node is globally N+1 redundant
Redundancy for DRBD nodes is already checked at every node locally.
However, in the presence of shared/external storage, we also have to
consider planning capacity for those instances---globally (within
a node group) as they can be moved to any node. Add a function that
checks whether this is true for a given node in a given cluster.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>