Add a type for a collection of allocation solutions
In this way, we can compute and rank allocation solutions while still
having the opportunity to filter out later unwanted ones. The advantage
is that the filtering predicate is only evaluated for the best allocations
till a suitable one is found. This is useful if the predicate to filter
on is expensive to compute.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>