SSH file manager: properly name assertion function
So far, the (fake) SSH file manager had a couple of check
functions (which return "True" or "False") and a couple of
functions which throw an exception instead of returning
"False" (as in those cases more information for debugging
is neeede). However, it was not obvious from the function
name which behavior to expect. This patch renames all
functions which throw exceptions to "Assert*" and removes
superfluous 'self.assertTrue(*)" calls around them.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>